Convert TimelineElement to TypeScript
